* dev: (29 commits)
  [Push Notifications] Add test controller to send out notifications while testing
  commmissions in project status counts
  added stuff for Randa
  Move custom fields to different custom field groups
  closer dash - dashboard - refactored
  Small Bug fixes for setting the visibility of default work queue columns
  Target different deployment
  Updated map-box zoom in on home
  Updating mapbox zoom
  Add push notification service and notification-enabled user queries
  Update workflow
  added stuff for Randa
  added stuff for Randa
  remove references to columns created dynamically
  remove unused imports
  minor cleanup
  hotfix: add BR System User access to ProcessStepEventService
  Fix migration to add missing column before creating an index
  Deploy in k8s
  [IL SREC] Update how we're calculating total cost for the proposal
  ...

# Conflicts:
#	client/src/views/flow/schedule/components/Map.vue

+++ src/main/resources/db/migration/procedures/brs/dashboards/R__rpt_company_dashboard.sql
@@ -19,7 +19,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -45,7 +45,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -73,7 +73,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -99,7 +99,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -125,7 +125,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -151,7 +151,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -177,7 +177,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -203,7 +203,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -229,7 +229,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -255,7 +255,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -281,7 +281,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -307,7 +307,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -333,7 +333,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -359,7 +359,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -385,7 +385,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -411,7 +411,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -437,7 +437,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -463,7 +463,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -489,7 +489,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -515,7 +515,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -541,7 +541,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
@@ -567,7 +567,7 @@ BEGIN
                              display_order,
                              company_count,
                              case when trend_count is not null then
-                                    round((company_count - trend_count)/greatest(trend_count,1) * 100) end as trend_count,
+                                    round((company_count::numeric - trend_count::numeric)/greatest(trend_count::numeric,1) * 100)::numeric end as trend_count,
                              major_milestone
 
                       from (
